minigrid-rl-llm/requirements.txt
matplotlib
minigrid
numpy
PyYAML
tensorboardX
torch
torch_ac
pydantic

